Rename integration test namespaces for Integer and Long overloads to improve organization
